LPD-29194 Remove unncesary poshi tests
LPS-183056. Confirm that the "Multiple" and "Include" radio buttons are deselected
LPS-176187. Confirm that the Selection option has a tooltip with the text message.
LPS-183056. Confirm the change of the Source Options from Cars to Names.
LPS-176187. Confirm that the Filter has been created and has the Name and Field Name fields
LPS-183056. Confirm that there is only one option of car in Custom Picklist / Preselected.